Update contacto.html
We add the new meta tags and icon for the page. For now I'm using the "footer.css" for give the new design to my page. the new typography that I'm using is from google fonts this is the link: <link href="https://fonts.googleapis.com/css?family=Crimson+Text" rel="stylesheet">.
The new navbar are using the #ffffff; color just for this template, I think it's viewing cool.
the class: .first_amlshow is in footer.css it's using the font-family crimson text from google fonts.
the class: mouseover_contacto is only for java events with the addClass .the_back1, so when somebody drag the mouse over the class it will show the background color grey! we set "none" to text-decoration for better view.
I set a border-top to the tag "hr" just to view suitabling of colors.
I'm using new color in the footer the new one is #111111;
I add the new icons in the footer that are ruby, jquery icon, html5 and CSS3 and the version of the page.